Refactor out minor component compositions into a new format called renderers
Renderers allow smaller clusters of components to be grouped together inside of a function that is external to the body of the page function itself.

Externalizing these components make them easier to test as their individual units, and then test how they are used when it comes time to do integration tests.